+++ cmd/urldecode/urldecode.go
@@ -1,7 +1,7 @@
 //
 // urldecode.go is a simple command line utility to decode a string in a URL friendly way.
 //
-// Copyright (c) 2018, Caltech
+// Copyright (c) 2019, Caltech
 // All rights not granted herein are expressly reserved by Caltech.
 //
 // Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met: